Improve translation of Trac users to GitHub
Existing code was confused about whether it wanted to use the usernames
or emails and used different keys when looking up the assignee and the
comment author, fix it to always use Trac usernames which seems more
logical and avoids an unnecessary database lookup.

Also obfuscate the emails if we do use it and also include Trac
username in the string used on GitHub to allow searching by it later.